+++ Matrix.cpp
@@ -1,10 +1,10 @@
 #include "Matrix.h"
-size_t Matrix::size()
+int Matrix::size()
 {
 	return this->cols * this->rows;
 }
 
-size_t Matrix::size() const
+int Matrix::size() const
 {
 	return this->cols * this->rows;
 }
