+++ Matrix1d.h
@@ -8,12 +8,13 @@ public:
 
 	Matrix1d(int n);
 	Matrix1d(const Matrix1d& M);
-	Matrix1d(long double* arr, size_t n);
+	Matrix1d(long double* arr, int n);
 
 	void fill(long double value) override;
+
 	int indexOf_max();
 	Matrix1d abs();
-	Matrix1d trunc(size_t from, size_t to);
+	Matrix1d trunc(int from, int to);
 
 	Matrix1d operator= (const Matrix1d& M);
 	Matrix1d operator- (const Matrix1d& M) const;
