+++ Matrix2d.h
@@ -7,7 +7,7 @@ class Matrix2d : public Matrix
 public:
 	long double** matrix;
 
-	Matrix2d(size_t n);
+	Matrix2d(int n);
 	Matrix2d(const Matrix2d& M);
 
 	void fill(long double value) override;
@@ -18,8 +18,8 @@ public:
 	Matrix1d column(int index);
 	Matrix1d row(int index);
 
-	void swapColumns(int row1, int row2, size_t from, size_t to);
-	void swapRows(int col1, int col2, size_t from, size_t to);
+	void swapColumns(int row1, int row2, int from, int to);
+	void swapRows(int col1, int col2, int from, int to);
 
 	Matrix2d operator= (const Matrix2d& M);
 	Matrix1d operator* (const Matrix1d& M) const;
