+++ PointArray.cpp
@@ -1,6 +1,6 @@
 #include "PointArray.h"
 
-PointArray::PointArray(size_t n)
+PointArray::PointArray(int n)
 {
 	this->length = n;
 	this->arr = new Point[n];
@@ -8,7 +8,7 @@ PointArray::PointArray(size_t n)
 
 PointArray::PointArray(const PointArray& pointArray) : PointArray(pointArray.getLength())
 {
-	for (size_t i = 0; i < this->length; ++i) {
+	for (int i = 0; i < this->length; ++i) {
 		this->arr[i] = pointArray.arr[i];
 	}
 }
@@ -18,12 +18,12 @@ PointArray::~PointArray()
 	delete arr;
 }
 
-size_t PointArray::getLength() const
+int PointArray::getLength() const
 {
 	return this->length;
 }
 
-PointArray PointArray::trunc(size_t begin, size_t end)
+PointArray PointArray::trunc(int begin, int end)
 {
 	PointArray truncatedArray = PointArray(end - begin);
 	for (int i = begin; i < end; ++i) {
